requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.7
apache-airflow==2.7.3
python-dotenv==1.0.0